# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EduManage School SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schools/{schoolId}/teachers:
    get:
      operationId: listTeachers
      tags: [teachers]
      summary: List teachers
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Teachers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teacher"
    post:
      operationId: createTeacher
      tags: [teachers]
      summary: Add a teacher
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeacherBody"
      responses:
        "201":
          description: Teacher created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
  /schools/{schoolId}/teachers/{teacherId}/status:
    patch:
      operationId: updateTeacherStatus
      tags: [teachers]
      summary: Activate or deactivate a teacher
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: teacherId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"

  # SCHOOL SETTINGS
  /schools/{schoolId}/settings:
    get:
      operationId: getSchoolSettings
      tags: [settings]
      summary: Get school settings
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: School settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolSettings"
    put:
      operationId: updateSchoolSettings
      tags: [settings]
      summary: Update school settings
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSchoolSettingsBody"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolSettings"

  # FEATURE TOGGLES
  /schools/{schoolId}/features:
    get:
      operationId: getFeatureToggles
      tags: [settings]
      summary: Get feature toggles
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Feature toggles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureToggles"
    put:
      operationId: updateFeatureToggles
      tags: [settings]
      summary: Update feature toggles
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeatureToggles"
      responses:
        "200":
          description: Updated feature toggles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureToggles"

  # REPORTS
...
components:
  schemas:
...
    Teacher:
      type: object
      properties:
        id:
          type: integer
        schoolId:
          type: integer
        name:
          type: string
        subject:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
      required: [id, schoolId, name, status, createdAt]
    CreateTeacherBody:
      type: object
      properties:
        name:
          type: string
        subject:
          type: string
        phone:
          type: string
        email:
          type: string
      required: [name]
    UpdateStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [active, inactive]
      required: [status]
    SchoolSettings:
      type: object
      properties:
        schoolId:
          type: integer
        schoolName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        academicYear:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        themeColor:
          type: string
          nullable: true
      required: [schoolId, schoolName, contactEmail]
    UpdateSchoolSettingsBody:
      type: object
      properties:
        schoolName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        address:
          type: string
        academicYear:
          type: string
        themeColor:
          type: string
    FeatureToggles:
      type: object
      properties:
        schoolId:
          type: integer
        attendanceEnabled:
          type: boolean
        financeEnabled:
          type: boolean
        salesEnabled:
          type: boolean
        reportsEnabled:
          type: boolean
        busEnabled:
          type: boolean
      required: [schoolId, attendanceEnabled, financeEnabled, salesEnabled, reportsEnabled, busEnabled]
...